LPS-105162 move 2_1_0 to 2_2_0
